+++ app/controllers/editor/projects_controller.rb
@@ -60,16 +60,25 @@ class Editor::ProjectsController < ApplicationController
   end
 
   def project_params
-    params.require(:project).permit(
-      :name, :slug, :description, :disable_all_emails,
-      :hide_values_on_pdfs,
-      :randomizations_enabled, :adverse_events_enabled, :blinding_enabled,
-      :handoffs_enabled, :auto_lock_sheets,
-      # Uploaded Logo
-      :logo, :logo_uploaded_at, :logo_cache, :remove_logo,
-      # Will automatically generate a site if the project has no site
-      :site_name
-    )
+    if @project.unblinded?(current_user)
+      params.require(:project).permit(
+        :name, :slug, :description, :disable_all_emails,
+        :hide_values_on_pdfs,
+        :randomizations_enabled, :adverse_events_enabled, :blinding_enabled,
+        :handoffs_enabled, :auto_lock_sheets,
+        # Uploaded Logo
+        :logo, :logo_uploaded_at, :logo_cache, :remove_logo
+      )
+    else
+      params.require(:project).permit(
+        :name, :slug, :description, :disable_all_emails,
+        :hide_values_on_pdfs,
+        :randomizations_enabled, :adverse_events_enabled,
+        :handoffs_enabled, :auto_lock_sheets,
+        # Uploaded Logo
+        :logo, :logo_uploaded_at, :logo_cache, :remove_logo
+      )
+    end
   end
 
   def create_member_invite